requirements.txt
numpy
pillow
opencv-python
scipy
matplotlib
pandas
